requirements.txt
Flask==3.0.3
pyodbc==5.1.0
azure-storage-blob==12.19.1
python-dotenv==1.0.1